refactor : 이미지에 파일 업로드 이벤트 넣는 함수 추상화 및 포스트 수정페이지 & 포스트 작성페이지에 적용
