+++ src/pages/post-creation/index.tsx
@@ -7,6 +7,7 @@ import { createPost } from '@/apis/postApis'
 import { getChannelInform } from '@/apis/channelApis'
 import FileUploadIcon from '@/assets/icons/fileUpload.svg'
 import handleImageFormData from '@/utils/handleImageFormData'
+import onGetImageFile from '@/utils/onGetImageFile'
 
 const breadOptions = ['조리빵', '특수빵', '식빵', '과자빵']
 
@@ -42,7 +43,7 @@ const PostCreation = () => {
 
   const [title, setTitle] = useState<string>('')
   const [detail, setDetail] = useState<string>('')
-  const [file, setFile] = useState<File | null>(null)
+  const [file, setFile] = useState<File | undefined>()
   const [selectedBread, setSelectedBread] = useState<BreadType>(null)
 
   const onClickEnrollPost = async () => {
@@ -60,21 +61,6 @@ const PostCreation = () => {
     const parsedToken = JSON.parse(token as string)
     await createPost(parsedToken, formData)
     navigate('/')
-  }
-
-  const onClickUploadImage = () => {
-    const fileInput = document.createElement('input')
-    fileInput.type = 'file'
-    fileInput.accept = 'image/gif, image/jpeg,image/png,image/jpg'
-
-    fileInput.onchange = (e: Event) => {
-      if (!e.target) return
-      const target = e.target as HTMLInputElement
-      if (!target.files || !target.files[0]) return
-
-      setFile(target.files[0])
-    }
-    fileInput.click()
   }
 
   return (
@@ -95,7 +81,7 @@ const PostCreation = () => {
           </p>
           <button
             className="pb-2"
-            onClick={onClickUploadImage}>
+            onClick={() => onGetImageFile((newImage) => setFile(newImage))}>
             <img src={FileUploadIcon} />
           </button>
         </PostTitleImageInputWrapper>
